requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
openai>=1.66.5
streamlit>=1.31.0
sqlite-vec>=0.1.6 